Add calculateCombination func

+++ src/main/scala/Utils/ClinksCalculator.scala
@@ -25,5 +25,5 @@ object ClinksCalculator:
     * @return n choose k
     */
   // TODO - Part 1 Step 1
-  def calculateCombination(n: Int, k: Int): Int = factorial(n) / (factorial(k) * factorial(n - k))
+  def calculateCombination(n: Int, k: Int): Int = (factorial(n) / (factorial(k) * factorial(n - k))).toInt
 end ClinksCalculator
